+++ tclNote_unicodeBlockList.tcl
@@ -1,0 +1,17 @@
+#tclNote
+#tclNote_unicodeBlockList.tcl
+##===================================================================
+#	Copyright (c) 2018 Yuji SODE <yuji.sode@gmail.com>
+#
+#	This software is released under the MIT License.
+#	See LICENSE or http://opensource.org/licenses/mit-license.php
+##===================================================================
+#A list of the Unicode block names and their ranges for tclNote
+##===================================================================
+#Unicode is a registered trademark of Unicode, Inc. in the United States and other countries.
+#Reference: Unicode, Inc. 1991-2018. the Unicode Consortium. derived on [2017-08-19 and 2018-08-01] and from: http://www.unicode.org/
+#
+##=== list ===
+#block names and their ranges (the Unicode Standard Version 10.0.0-11.0.0)
+	source -encoding utf-8 ./tclNote_unicodeBlock.tcl;
+##===================================================================